Factor out API profile tests into module
